Support triggering by usernames that contain a hyphen
Previously, if the trigger comment was made by a user with a hyphen in their user name, the script errors.

For example, if the user name was "foo-bar":
```
/entrypoint.sh: line 57: foo-bar_TOKEN: invalid variable name
```
Bash does not permit hyphens in variable names, but GitHub does permit hyphens in user names.

The fix is to do a character replacement of any hyphens in the user name before using it as a component of the variable name.

This will require that the workflow maintainer also make this character replacement when configuring user-specific tokens in the workflow for users with hyphens in their user name. Since there is no documentation of the user-specific token feature of this action, I was not able to update it to note this requirement.