Bump `@1.4` to `@1.5` in README
Was trying to get `env: PR_NUMBER: xxx` to work via copying the code in the README. Took me a while to realize that y'all fixed it in `cirrus-actions/rebase@1.5`! Just bumping the sample code up a version so that it just works upon copy/paste 😄 

Thanks for making this! ❤️